refactor(server): peel leftover singleton routes into routers/misc.py (router-split, round-5 #51 / R5-25 PR23)
Twelfth router peel. Moves the standalone handlers that were keeping server.py
from being a pure app shell into a leaf routers/misc.py: GET /api/stream
(range-friendly playback + H.264 proxy fallback for HEVC/ProRes), POST
/api/embed/rebuild (single-flight semantic-index rebuild), POST /api/open-file
(OS open/reveal, videos_write-gated), and the unauthenticated POST /api/client-log
diagnostics sink — with _log_safe (its only caller) and the OpenFileRequest /
ClientLogRequest models.

Shared deps imported from leaf modules: _proxy_ready + _resolve_media_path from
pathres, the embed single-flight guard + _rebuild_embeddings worker from state
(the ONE shared instance), _assert_same_site from webguard; MEDIA_EXTS via
mediatypes.MEDIA_EXT directly. The app shell — SPA "/" (serve_index/_load_index),
/thumbnails, lifespan, middleware, include_router wiring — stays in server.py.

Behaviour-preserving: paths/methods/scopes unchanged (SPA + CLI unaffected).
server.py 2682 → 2516. test_server _log_safe test repointed to routers.misc;
export-builders + proxy boundary tests updated (helpers/stream_media relocated).
