+++ tests/test_r5_25_export_builders_module.py
@@ -50,12 +50,19 @@ def test_server_reexports_export_builders_by_identity():
         )
 
 
-def test_http_and_log_helpers_stay_in_server():
+def test_http_and_log_helpers_did_not_go_to_export_builders():
     # _attachment_headers (Content-Disposition) and _log_safe (terminal sanitiser)
-    # are HTTP/logging concerns, not export-format serialisers — they must NOT move.
+    # are HTTP/logging concerns, not export-format serialisers — they must NOT have
+    # been pulled into export_builders. _log_safe later moved to routers/misc.py
+    # with its only caller /api/client-log (R5-25 misc peel); _attachment_headers
+    # still lives in server (moves with the export routes in a later peel).
+    import export_builders
+    assert not hasattr(export_builders, "_attachment_headers")
+    assert not hasattr(export_builders, "_log_safe")
     import server
     assert hasattr(server, "_attachment_headers")
-    assert hasattr(server, "_log_safe")
+    import routers.misc
+    assert hasattr(routers.misc, "_log_safe")
 
 
 # ── injection-hardening / format behaviour preserved by the move ─────────────
