+++ tests/test_server.py
@@ -1279,12 +1279,13 @@ def test_log_safe_strips_control_chars_and_truncates():
     """client-log fields must not carry newlines/ANSI escapes (log injection) or
     be unbounded (disk fill)."""
     import importlib
-    server = importlib.import_module("server")
-    out = server._log_safe("hello\nFAKE LOG\x1b[31m evil\r\n", 100)
+    # _log_safe moved to routers/misc.py with /api/client-log in the R5-25 split.
+    misc = importlib.import_module("routers.misc")
+    out = misc._log_safe("hello\nFAKE LOG\x1b[31m evil\r\n", 100)
     assert "\n" not in out and "\r" not in out and "\x1b" not in out
     assert "hello" in out and "evil" in out
-    assert server._log_safe("x" * 5000, 16) == "x" * 16
-    assert server._log_safe("音樂 OK", 100) == "音樂 OK"  # CJK preserved
+    assert misc._log_safe("x" * 5000, 16) == "x" * 16
+    assert misc._log_safe("音樂 OK", 100) == "音樂 OK"  # CJK preserved
 
 
 # ── Track A hardening: input validation ──────────────────────────────────────
